allow exporting spaces
