+++ apps/cron/src/tasks/exportSpaceData/index.ts
@@ -1,0 +1,1 @@
+export * from './exportSpaceDataTask';
