--- apps/webapp/components/common/CharmEditor/components/nestedPage/nestedPage.constants.ts
@@ -1,2 +1,0 @@
-export const nestedPageNodeName = 'page';
-export const nestedPageMarkdownEnclosure = 'MARKDOWN_NESTED_PAGE';
