+++ apps/webapp/components/common/PageActions/utils/exportMarkdown.ts
@@ -1,7 +1,6 @@
 import type { Prisma } from '@charmverse/core/prisma';
-
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
 import type { Member } from '@packages/lib/members/interfaces';
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
 
 import { downloadMarkdownFile } from './downloadMarkdownFile';
