+++ apps/webapp/components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/components/VoteEvaluation/components/PublishToSnapshot/components/PublishingForm.tsx
@@ -6,7 +6,12 @@ import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
 import Tooltip from '@mui/material/Tooltip';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
 import { getChainById } from '@packages/blockchain/connectors/chains';
+import { getSnapshotClient } from '@packages/lib/snapshot/getSnapshotClient';
+import { getSnapshotSpace } from '@packages/lib/snapshot/getSpace';
+import type { SnapshotReceipt, SnapshotSpace, SnapshotVotingStrategy } from '@packages/lib/snapshot/interfaces';
+import { SnapshotVotingMode } from '@packages/lib/snapshot/interfaces';
 import { ExternalServiceError, MissingWeb3AccountError, SystemError, UnknownError } from '@packages/utils/errors';
 import { lowerCaseEqual } from '@packages/utils/strings';
 import type { ProposalType } from '@snapshot-labs/snapshot.js/dist/sign/types';
@@ -27,11 +32,6 @@ import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useIsAdmin } from 'hooks/useIsAdmin';
 import { useMembers } from 'hooks/useMembers';
 import { useWeb3Account } from 'hooks/useWeb3Account';
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
-import { getSnapshotClient } from '@packages/lib/snapshot/getSnapshotClient';
-import { getSnapshotSpace } from '@packages/lib/snapshot/getSpace';
-import type { SnapshotReceipt, SnapshotSpace, SnapshotVotingStrategy } from '@packages/lib/snapshot/interfaces';
-import { SnapshotVotingMode } from '@packages/lib/snapshot/interfaces';
 
 import ConnectSnapshot from './ConnectSnapshot';
 import { InputVotingStrategies } from './InputVotingStrategies';
