+++ apps/webapp/hooks/useCreateLensPublication.ts
@@ -2,6 +2,7 @@ import { log } from '@charmverse/core/log';
 import { textOnly } from '@lens-protocol/metadata';
 import type { Session } from '@lens-protocol/react-web';
 import { SessionType, useCreateComment, useCreatePost } from '@lens-protocol/react-web';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
 import type { PageContent } from '@packages/charmeditor/interfaces';
 import { switchActiveNetwork } from '@packages/lib/blockchain/switchNetwork';
 import { LensChain } from '@packages/lib/lens/constants';
@@ -13,7 +14,6 @@ import { useHandleLensError, useLensProfile } from 'components/settings/account/
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useSnackbar } from 'hooks/useSnackbar';
 import { useWeb3Account } from 'hooks/useWeb3Account';
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
 
 import { usePage } from './usePage';
