+++ apps/webapp/lib/pages/comments/syncPageCommentsWithLensPost.ts
@@ -1,7 +1,6 @@
 import { Prisma, prisma } from '@charmverse/core/prisma-client';
+import { parseMarkdown } from '@packages/bangleeditor/markdown/parseMarkdown';
 import { fetchLensPageComments } from '@packages/lib/lens/fetchLensPostComments';
-
-import { parseMarkdown } from 'lib/prosemirror/markdown/parseMarkdown';
 
 import { listPageComments } from './listPageComments';
