+++ apps/webapp/lib/pages/getPageMarkdown.ts
@@ -1,6 +1,7 @@
 import { prisma } from '@charmverse/core/prisma-client';
 import { _ } from '@packages/bangleeditor/builders';
 import { getNodeFromJson } from '@packages/bangleeditor/getNodeFromJson';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
 import { getChainById } from '@packages/blockchain/connectors/chains';
 import type {
   ProjectField,
@@ -16,8 +17,6 @@ import type {
   SelectOptionType
 } from '@packages/lib/proposals/forms/interfaces';
 import type { ProposalFields } from '@packages/lib/proposals/interfaces';
-
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
 
 export async function getPageMarkdown({
   pageId,
