+++ apps/webapp/lib/pages/interfaces.ts
@@ -27,11 +27,6 @@ export interface TrashOrDeletePageResponse {
   pageIds: string[];
 }
 
-export interface PageLink {
-  title: string;
-  url: string;
-}
-
 // extend PageMeta so that we can populate usePages hook while /pages is loading
 export type PageWithContent = PageMeta &
   Pick<
