+++ apps/webapp/lib/pages/server/index.ts
@@ -1,4 +1,3 @@
 export * from './getPage';
 export * from '@packages/pages/errors';
-export * from './generatePageLink';
 export * from '../interfaces';
