+++ apps/webapp/lib/prosemirror/getMarkdownText.ts
@@ -1,6 +1,5 @@
 import type { Prisma } from '@charmverse/core/prisma';
-
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
 
 export async function getMarkdownText(content: Prisma.JsonValue | null) {
   try {
