+++ apps/webapp/pages/api/pages/[id]/link.ts
@@ -2,11 +2,11 @@ import type { Page, Space } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 import { onError, onNoMatch, requireUser } from '@packages/lib/middleware';
 import { withSessionRoute } from '@packages/lib/session/withSession';
+import { PageNotFoundError } from '@packages/pages/errors';
+import { generatePageLink } from '@packages/pages/generatePageLink';
 import { hasAccessToSpace } from '@packages/users/hasAccessToSpace';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
-
-import { generatePageLink, PageNotFoundError } from 'lib/pages/server';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
