+++ apps/webapp/pages/api/v1/forum/comments/[commentId]/upvote.ts
@@ -1,14 +1,14 @@
 import { UnauthorisedActionError } from '@charmverse/core/errors';
 import { prisma } from '@charmverse/core/prisma-client';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
+import { voteForumComment } from '@packages/lib/forums/posts/voteForumComment';
+import { requireKeys } from '@packages/lib/middleware';
+import { withSessionRoute } from '@packages/lib/session/withSession';
 import { InvalidInputError } from '@packages/utils/errors';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
-import { voteForumComment } from '@packages/lib/forums/posts/voteForumComment';
-import { requireKeys } from '@packages/lib/middleware';
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
 import { superApiHandler } from 'lib/public-api/handler';
 import { userProfileSelect, getUserProfile } from 'lib/public-api/searchUserProfile';
-import { withSessionRoute } from '@packages/lib/session/withSession';
 
 import type { PublicApiPostComment } from '../../posts/[postId]/comments/index';
