+++ apps/webapp/pages/api/v1/proposals/index.ts
@@ -1,13 +1,13 @@
 import type { ProposalEvaluationType, ProposalStatus } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 import { getCurrentEvaluation } from '@charmverse/core/proposals';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
+import type { ProposalEvaluationResultExtended } from '@packages/lib/proposals/interfaces';
+import { withSessionRoute } from '@packages/lib/session/withSession';
 import { InvalidStateError } from '@packages/nextjs/errors';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
-import type { ProposalEvaluationResultExtended } from '@packages/lib/proposals/interfaces';
-import { generateMarkdown } from 'lib/prosemirror/markdown/generateMarkdown';
 import { apiHandler } from 'lib/public-api/handler';
-import { withSessionRoute } from '@packages/lib/session/withSession';
 
 const handler = apiHandler();
