+++ apps/webapp/tsconfig.json
@@ -46,5 +46,5 @@
     "types": ["vitest/globals"]
   },
   "exclude": ["node_modules", "scripts"],
-  "include": ["**/*.ts", "**/*.tsx", "../../packages/pages/src/errors.ts", "../../packages/lib/src/rewards/blocks/getDefaultRewardProperties.ts"]
+  "include": ["**/*.ts", "**/*.tsx"]
 }
