+++ packages/spaces/package.json
@@ -11,6 +11,8 @@
     "./*": "./src/*.ts"
   },
   "dependencies": {
+    "@packages/databases": "^0.0.0",
+    "@packages/lib": "^0.0.0",
     "@packages/utils": "^1.0.0",
     "viem": "^2.29.2"
   }
