+++ packages/spaces/src/exportSpaceData.ts
@@ -1,0 +1,62 @@
+import { log } from '@charmverse/core/log';
+import { generateMarkdown } from '@packages/bangleeditor/markdown/generateMarkdown';
+import { onError, onNoMatch } from '@packages/lib/middleware';
+import { permissionsApiClient } from '@packages/lib/permissions/api/client';
+import { withSessionRoute } from '@packages/lib/session/withSession';
+import type { ContentToCompress, MarkdownPageToCompress } from '@packages/lib/utils/file';
+import { zipContent } from '@packages/lib/utils/file';
+import { paginatedPrismaTask } from '@packages/lib/utils/paginatedPrismaTask';
+import type { NextApiRequest, NextApiResponse } from 'next';
+import nc from 'next-connect';
+
+const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
+
+handler.post(requestZip);
+
+export type ZippedDataRequest = Pick<ContentToCompress, 'csv'> & { pageIds: string[] };
+
+export async function requestZip({ userId, spaceId }: { userId: string; spaceId: string }) {
+  const accessiblePageIds = await permissionsApiClient.pages.getAccessiblePageIds({
+    spaceId,
+    userId,
+    archived: false
+    // limit,
+    // filter,
+    // search
+  });
+
+  const markdownPages = await paginatedPrismaTask({
+    model: 'page',
+    mapper: async (page) => {
+      const markdown = await generateMarkdown({
+        content: page.content
+      }).catch((err) => {
+        log.error('Error generating markdown', { pageId: page.id, error: err });
+        return 'Error generating markdown for this page';
+      });
+
+      return {
+        title: page.title || 'Untitled',
+        contentMarkdown: markdown ?? ''
+      } as MarkdownPageToCompress;
+    },
+    queryOptions: {
+      where: {
+        id: {
+          in: accessiblePageIds
+        }
+      },
+      select: {
+        id: true,
+        title: true,
+        content: true
+      }
+    }
+  });
+
+  const compressed = await zipContent({ csv: req.body.csv ?? [], pages: markdownPages });
+
+  return res.status(200).setHeader('Content-Type', 'application/octet-stream').send(compressed);
+}
+
+export default withSessionRoute(handler);
